PROVISORY: ci: cram: obuspa: whitelist missing Device.Boot! event arguments
tr181-device registers the event path "Device.Boot!" but does not
register any parameters associated with the event.

See https://usp-data-models.broadband-forum.org/tr-181-2-18-1-usp.html#D.Device:2.Device.Boot

The reason we didn't see this before, is because the obuspa 8.0.7 was
handling the Device.Boot! event internally.  This had to be handled by
obuspa as a special case because Device.Boot! is part of the Device.
object and could not be registered by a USP service in that version.

obuspa 9.0.3 now expects tr181-device service to register Device.Boot!
as a (non-object path) event.

tr181-device registers the event path "Device.Boot!" but does not
register any parameters associated with the event. We can see this from
the GET_SUPPORTED_DM_RESP message from tr181-device to obuspa (you can
obtain this by adding -v to the broker cmd line).

The current event in the tr181-device_definition.odl is:

  event "Boot!";

but should contain:

  event "Boot!" {
    string Cause;
    string CommandKey;
    bool FirmwareUpdated;
    string ParameterMap;
  }

So lets whitelist the missing event arguments till PCF-1529 is handled,
so this test workaround can be removed.

References: PCF-1529
Reported-by: Steven Corby <steven.corby@vantiva.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>