scripts: gen_config: allow to remove a package
The 'diffconfig' section was parsed before the packages sections.
Therefore it was not possible to remove a package that was once added in
a profile, in another, subsequent profile.
Like so for example:

    ./scripts/gen_config.py prpl override

Where the content of override.yml is:

diffconfig: |
      # CONFIG_PACKAGE_cthulhu-lxc is not set

Fix this by parsing diffconfig section after packages sections.

Fixes: PCF-1587
Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>