fstools: partname: drop use of parent_dev
partname_volume_init() calls block_volume_format() passing the device
path.

Previously the parent path was passed typically leading to retrieve the
size of the parent device and not the device to be formatted.
Eventually, formatting the block device with mkfs.f2fs failed silently
when the block device is too small and created a ram overlayfs instead
of trying to format with ext4.

Patch was sent upstream for review in
https://patchwork.ozlabs.org/project/openwrt/patch/tencent_1BDE4F7AE014AD4D4FA835C5DFEAB35AFB06@qq.com/

Fixes: PCF-1893
Tracked-by: FEAT-29
Signed-off-by: Paul HENRYS <paul.henrys_ext@softathome.com>